# Conflicts:
#	src/main/java/com/mci/ticketpilot/views/DashboardView.java
#	src/main/java/com/mci/ticketpilot/views/MainLayout.java
#	src/main/java/com/mci/ticketpilot/views/lists/TicketForm.java

+++ src/main/java/com/mci/ticketpilot/data/service/PilotService.java
@@ -85,7 +85,11 @@ public class PilotService {
     public List<Project> findAllProjects(){ return projectRepository.findAll(); }
     public long countProjects() { return projectRepository.count(); }
     public void deleteProject(Project project) {
-        projectRepository.delete(project);
+        if(project.getTickets() == null || project.getTickets().isEmpty()) {
+            projectRepository.delete(project);
+        } else {
+            System.err.println("Project has tickets. Cannot delete.");
+        }
     }
 
     public void saveProject(Project project) {
